Changed components of login page to tailwind css
