--- src/app/auth/login/loginpage.css
@@ -1,101 +1,0 @@
-.loginLeft {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    padding: 32px 0px 64px;
-    gap: 32px;
-    width: 464px;
-    height: 716px;
-    flex: none;
-    order: 1;
-    flex-grow: 0;
-  }
-  
-  .loginLeft h1 {
-    font-size: 36px;
-    font-weight: 700;
-    line-height: 44px;
-    letter-spacing: 0em;
-    text-align: left;
-  }
-  
-  .loginLeft span {
-    font-size: 16px;
-    font-weight: 400;
-    line-height: 24px;
-    letter-spacing: 0em;
-    text-align: left;
-  }
-  
-  .loginLeft .max-w-4 {
-    max-width: 464px;
-  }
-  
-  .loginLeft form {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    padding: 0px;
-    gap: 16px;
-    width: 100%;
-  }
-  
-  .loginLeft form input {
-    box-sizing: border-box;
-    width: 100%;
-    height: 48px;
-    background: #FFFFFF;
-    border: 1px solid #D6D9E0;
-    border-radius: 8px;
-    padding: 12px 16px;
-    font-size: 16px;
-    font-weight: 400;
-    line-height: 24px;
-    letter-spacing: 0em;
-    text-align: left;
-  }
-  
-  .loginLeft form .remember-me {
-    display: flex;
-    align-items: center;
-    width: 100%;
-  }
-  
-  .loginLeft form .remember-me input[type="checkbox"] {
-    width: 16px;
-    height: 16px;
-    margin-right: 8px;
-  }
-  
-  .loginLeft form .remember-me label {
-    font-size: 14px;
-    font-weight: 400;
-    line-height: 20px;
-    letter-spacing: 0em;
-    text-align: left;
-  }
-  
-  .loginLeft form .forgot-password {
-    font-size: 14px;
-    font-weight: 400;
-    line-height: 20px;
-    letter-spacing: 0em;
-    text-align: right;
-    width: 100%;
-    color: #3B82F6;
-    cursor: pointer;
-  }
-  
-  .loginLeft form button {
-    width: 100%;
-    height: 48px;
-    background: #3B82F6;
-    border-radius: 8px;
-    font-size: 16px;
-    font-weight: 600;
-    line-height: 24px;
-    letter-spacing: 0em;
-    text-align: center;
-    color: #FFFFFF;
-    cursor: pointer;
-  }
